Refactor pokeapi.go, add Nature and Stat structs, implement GetNature and GetStat functions
- Added Nature struct with Name and ID fields
- Added Stat struct with Name and ID fields
- Implemented GetNature function to retrieve a nature by name or id
- Implemented GetStat function to retrieve a stat by name or id
